backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Stripe Payments
stripe==8.2.0

# AI & ML
anthropic==0.18.1
langchain==0.1.9
langchain-anthropic==0.1.4
langgraph==0.0.26
chromadb==0.4.22
openai==1.12.0
tiktoken==0.6.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
pillow==10.2.0
pytesseract==0.3.10

# Utilities
httpx==0.26.0
pydantic-core==2.16.2
typing-extensions==4.9.0
email-validator==2.1.0
tenacity==8.2.3

# Development
pytest==8.0.0
pytest-asyncio==0.23.4
black==24.2.0
ruff==0.2.2
